+++ authing/test/authing_test_auto.dart
@@ -1,13 +1,20 @@
+import 'package:flutter_test/flutter_test.dart';
+
 import 'package:authing_sdk/authing.dart';
 import 'package:authing_sdk/client.dart';
 import 'package:authing_sdk/oidc/oidc_client.dart';
 import 'package:authing_sdk/result.dart';
 import 'package:authing_sdk/user.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
 // can run all case serially in one go
 
 void main() {
+
+  // on mac, when running test, it will crash without this line
+  SharedPreferences.setMockInitialValues({});
+
   String pool = "60caaf41da89f1954875cee1";
   String appid = "60caaf41df670b771fd08937";
   Authing.init(pool, appid);
@@ -36,8 +43,7 @@ void main() {
   });
 
   test('register by username', () async {
-    AuthResult result =
-        await AuthClient.registerByUserName("test1024", "111111");
+    AuthResult result = await AuthClient.registerByUserName("test1024", "111111");
     expect(result.code, 200);
     expect(result.user?.username, "test1024");
     expect(result.user?.token != null, true);
@@ -46,8 +52,7 @@ void main() {
     expect(result2.code, 200);
     expect(result2.user?.username, "test1024");
 
-    AuthResult result3 =
-        await AuthClient.registerByUserName("test1024", "111111");
+    AuthResult result3 = await AuthClient.registerByUserName("test1024", "111111");
     expect(result3.code, 2026);
 
     Result result5 = await AuthClient.deleteAccount();
@@ -99,8 +104,7 @@ void main() {
     expect(result.code, 200);
     expect(result.user?.username, "ci");
 
-    AuthResult result2 =
-        await AuthClient.getCustomData(AuthClient.currentUser!.id);
+    AuthResult result2 = await AuthClient.getCustomData(AuthClient.currentUser!.id);
     expect(result2.code, 200);
     expect(AuthClient.currentUser?.customData[0]["key"], "org");
     expect(AuthClient.currentUser?.customData[0]["value"], "unit_test");
@@ -117,8 +121,7 @@ void main() {
     expect(AuthClient.currentUser?.customData[0]["value"], "unit_test");
 
     AuthClient.currentUser?.customData[0]["value"] = "hello";
-    AuthResult result3 =
-        await AuthClient.setCustomData(AuthClient.currentUser!.customData);
+    AuthResult result3 = await AuthClient.setCustomData(AuthClient.currentUser!.customData);
     expect(result3.code, 200);
     expect(AuthClient.currentUser?.customData[0]["value"], "hello");
 
@@ -131,13 +134,15 @@ void main() {
     expect(result.code, 200);
     expect(result.user?.username, "ci");
 
-    AuthResult result2 =
-        await AuthClient.updateProfile({"username": "hey", "nickname": "musk"});
+    AuthResult result2 = await AuthClient.updateProfile({
+      "username":"hey",
+      "nickname":"musk"
+    });
     expect(result2.code, 200);
     expect(result2.user?.username, "hey");
     expect(result2.user?.nickname, "musk");
 
-    AuthResult result3 = await AuthClient.updateProfile({"username": "ci"});
+    AuthResult result3 = await AuthClient.updateProfile({"username":"ci"});
     expect(result3.code, 200);
     expect(result3.user?.username, "ci");
   });